+++ src/test/specs/helpers.js
@@ -8,6 +8,8 @@ define([
 
     this.datasource = {};
     this.annotationsSrv = {};
+    this.timeSrv = new TimeSrvStub();
+    this.templateSrv = new TemplateSrvStub();
     this.datasourceSrv = {
       getMetricSources: function() {},
       get: function() { return self.datasource; }
@@ -17,6 +19,8 @@ define([
       return module(function($provide) {
         $provide.value('datasourceSrv', self.datasourceSrv);
         $provide.value('annotationsSrv', self.annotationsSrv);
+        $provide.value('timeSrv', self.timeSrv);
+        $provide.value('templateSrv', self.templateSrv);
       });
     };
 
@@ -25,7 +29,6 @@ define([
         self.scope = $rootScope.$new();
         self.scope.panel = {};
         self.scope.row = { panels:[] };
-        self.scope.filter = new FilterSrvStub();
         self.scope.dashboard = {};
         self.scope.dashboardViewState = new DashboardViewStateStub();
 
@@ -50,7 +53,6 @@ define([
         self.service = InfluxDatasource;
         self.$q = $q;
         self.$rootScope = $rootScope;
-        self.filterSrv = new FilterSrvStub();
         self.$httpBackend =  $httpBackend;
       }]);
     };
@@ -61,7 +63,7 @@ define([
     };
   }
 
-  function FilterSrvStub() {
+  function TimeSrvStub() {
     this.time = { from:'now-1h', to: 'now'};
     this.timeRange = function(parse) {
       if (parse === false) {
@@ -73,15 +75,21 @@ define([
       };
     };
 
-    this.applyTemplateToTarget = function(target) {
+    this.replace = function(target) {
       return target;
     };
   }
 
+  function TemplateSrvStub() {
+    this.variables = [];
+    this.replace = function() {};
+  }
+
+
 
   return {
     ControllerTestContext: ControllerTestContext,
-    FilterSrvStub: FilterSrvStub,
+    TimeSrvStub: TimeSrvStub,
     ServiceTestContext: ServiceTestContext
   };